reorganized types to there own module
