+++ crab-core/src/lib.rs
@@ -1,13 +1,7 @@
 use logging::trace;
-use null::Null;
-use std::{
-    collections::HashMap,
-    sync::{LazyLock, RwLock},
-};
 
-pub mod int;
-pub mod null;
-pub mod text;
+pub mod factory;
+pub mod types;
 
 /// The kinds of errors that can occur around objects
 pub enum ObjectError {
@@ -50,7 +44,7 @@ impl Key {
 }
 
 /// Extracts the key from a byte slice and returns what is remaining in the slice
-// TODO: Should this be an associated function
+// TODO: Should this be an associated function?
 pub fn extract_key(slice: &[u8]) -> Result<(Key, &[u8]), ObjectError> {
     trace!("Deserializing key");
 
@@ -211,92 +205,4 @@ impl From<DbObject> for ObjectData {
             data: object.into_raw_object_data(),
         }
     }
-}
-
-/// Responsible for creating an object
-#[derive(Debug)]
-pub struct ObjectFactory<F>
-where
-    F: Fn(RawObjectData) -> Result<Box<dyn Object>, ObjectError>,
-{
-    creator: F,
-}
-
-impl<F> ObjectFactory<F>
-where
-    F: Fn(RawObjectData) -> Result<Box<dyn Object>, ObjectError>,
-{
-    /// Creates a new object factory
-    pub fn new(creator: F) -> Self {
-        Self { creator }
-    }
-
-    /// Creates an object
-    pub fn create(&self, data: RawObjectData) -> Result<Box<dyn Object>, ObjectError> {
-        (self.creator)(data)
-    }
-}
-
-/// Just shortand for the ObjectFactory that the TypeRegistry uses
-type TypeRegistryFactoryType =
-    ObjectFactory<Box<dyn Fn(RawObjectData) -> Result<Box<dyn Object>, ObjectError> + Sync + Send>>;
-
-/// Responsible for holding and managing mappings of type ids to methods to create the types
-pub struct TypeRegistry {
-    registry: HashMap<ObjectType, TypeRegistryFactoryType>,
-}
-
-impl TypeRegistry {
-    /// Creates a new type registry
-    pub fn new() -> Self {
-        Self {
-            registry: HashMap::new(),
-        }
-    }
-
-    /// Inserts a factory into the registry with an associated type id
-    pub fn add_factory(&mut self, type_id: ObjectType, factory: TypeRegistryFactoryType) {
-        self.registry.insert(type_id, factory);
-    }
-
-    /// Gets a ObjectFactory from the registry
-    pub fn get_factory(&self, type_id: ObjectType) -> Option<&TypeRegistryFactoryType> {
-        self.registry.get(&type_id)
-    }
-}
-
-impl Default for TypeRegistry {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// There is only one type registry that must be shared with everything
-static REGISTRY: LazyLock<RwLock<TypeRegistry>> =
-    LazyLock::new(|| RwLock::new(TypeRegistry::default()));
-
-/// Adds a new ObjectFactory to the type registry
-pub fn register_factory(type_id: ObjectType, factory: TypeRegistryFactoryType) {
-    REGISTRY.write().unwrap().add_factory(type_id, factory);
-}
-
-/// Creates a new object and derives the type id from the input data
-pub fn new_db_object(object_data: ObjectData) -> Result<DbObject, ObjectError> {
-    // getting the factory
-    let registry = REGISTRY.read().unwrap();
-
-    let ObjectData { type_id, data } = object_data;
-
-    let Some(factory) = registry.get_factory(type_id) else {
-        return Err(ObjectError::MissingFactory);
-    };
-
-    let object = factory.create(data)?;
-
-    Ok(DbObject::new(type_id, object))
-}
-
-/// Creates a Null object
-pub fn new_null_db_object() -> DbObject {
-    DbObject::new(0, Box::new(Null))
 }
