+++ crab-core/src/types/int.rs
