+++ crab-core/src/types/text.rs
